Improve handling of style tags
The current implementation required <style> tags to be added within the control's HTML, basically meaning the <style>s were added every single time a control was created.

Now, we can put a <style> tag outside of the control's HTML, and WebControls will add it to the document's <head> tag, so it'll only appear once.